Renamed tribler client file
